+++ src/lightcode/op.cljs
@@ -3,15 +3,18 @@
    ["axios" :as axios]
 
    [kitchen-async.promise :as p]
-   [lightcode.lib :as lib]))
+   [lightcode.config :as config]
+   [lightcode.editor :as editor]
+   [lightcode.workspace :as workspace]))
 
-(def server-url
-  "http://localhost:8383/nrepl")
 
+(defn send! [env message]
+  (let [context {:context (merge {:nrepl {:port (editor/nrepl-port!)}
+                                  :env   env}
 
-(defn send! [message]
-  (let [message (clj->js (assoc message :port (lib/nrepl-port!)))]
-    (.post axios server-url message)))
+                                 (workspace/cljs-repl-context!))}
+        message (merge message context)]
+    (.post axios config/repl-api-url (clj->js message))))
 
 
 (defn ->info
@@ -24,12 +27,29 @@
 
 
 (defn info!
-  "Sends an info message.
+  "Sends an 'info' message.
   
-   `ns` Namespace name, e.g. lightcode.op
+  `ns` Namespace name, e.g. lightcode.op
+  `symbol` Symbol, e.g. send!"
+  [env ns symbol]
+  (send! env (->info ns symbol)))
+
+(defn ->eldoc
+  "`ns` Namespace name, e.g. lightcode.op
    `symbol` Symbol, e.g. send!"
   [ns symbol]
-  (send! (->info ns symbol)))
+  {:op     "eldoc"
+   :ns     ns
+   :symbol symbol})
+
+
+(defn eldoc!
+  "Sends an 'eldoc' message.
+  
+  `ns` Namespace name, e.g. lightcode.op
+  `symbol` Symbol, e.g. send!"
+  [env ns symbol]
+  (send! env (->eldoc ns symbol)))
 
 
 (defn ->load-file
@@ -46,5 +66,47 @@
   "`content` Full contents of a file of code
    `path` Source-path-relative path of the source file, e.g. clojure/java/io.clj
    `name` Name of source file, e.g. io.clj"
-  [content path name]
-  (send! (->load-file content path name)))
+  [env content path name]
+  (send! env (->load-file content path name)))
+
+
+(defn ->ns-vars
+  "`ns` Namespace name, e.g. lightcode.op"
+  [ns]
+  {:op "ns-vars"
+   :ns ns})
+
+
+(defn ns-vars!
+  "Sends a 'ns-vars' message.
+  
+  `ns` Namespace name, e.g. lightcode.op
+  `symbol` Symbol, e.g. send!"
+  [env ns]
+  (send! env (->ns-vars ns)))
+
+
+(defn ->ns-vars-with-meta
+  "`ns` Namespace name, e.g. lightcode.op"
+  [ns]
+  {:op "ns-vars-with-meta"
+   :ns ns})
+
+
+(defn ns-vars-with-meta!
+  "Sends a 'ns-vars-with-meta' message.
+  
+  `ns` Namespace name, e.g. lightcode.op
+  `symbol` Symbol, e.g. send!"
+  [env ns]
+  (send! env (->ns-vars-with-meta ns)))
+
+
+(defn ->ns-load-all []
+  {:op "ns-load-all"})
+
+
+(defn ns-load-all!
+  "Sends a 'ns-load-all' message."
+  [env]
+  (send! env (->ns-load-all)))